The client now immeditaly sends the objects to the object storage server.
